product_search/requirements.txt
python-amazon-simple-product-api==2.2.11
